requirements.txt
requests
python-dotenv
yt-dlp
apscheduler
fastapi
uvicorn
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
aiofiles